split publish handlers into components
