+++ apps/chat/src/components/Chat/Chat.tsx
@@ -74,7 +74,7 @@ import { MemoizedChatMessage } from './MemoizedChatMessage';
 import { NotAllowedModel } from './NotAllowedModel';
 import { PlaybackControls } from './Playback/PlaybackControls';
 import { PublicationControls } from './Publish/PublicationChatControls';
-import { PublicationHandler } from './Publish/PublicationHandler';
+import { PublicationHandler } from './Publish/PublicationHandler/PublicationHandler';
 import { TalkToModal } from './TalkTo/TalkToModal';
 
 import {
