+++ apps/chat/src/components/Chat/Publish/PublicationHandler/PublicationFilters.tsx
@@ -1,0 +1,57 @@
+import { useTranslation } from 'next-i18next';
+
+import { Publication, PublicationRule } from '@/src/types/publication';
+import { Translation } from '@/src/types/translation';
+
+import { Spinner } from '@/src/components/Common/Spinner';
+
+import { RuleListItem } from '../RuleListItem';
+
+interface FilterComponentProps {
+  filteredRuleEntries: [string, PublicationRule[]][];
+  newRules: PublicationRule[];
+  publication: Publication;
+  isRulesLoading: boolean;
+}
+
+export function PublicationFilters({
+  filteredRuleEntries,
+  newRules,
+  publication,
+  isRulesLoading,
+}: FilterComponentProps) {
+  const { t } = useTranslation(Translation.Chat);
+
+  if (isRulesLoading) {
+    return (
+      <div className="flex size-full items-center justify-center">
+        <Spinner size={48} />
+      </div>
+    );
+  }
+
+  const isNoRulesToDisplay =
+    (!filteredRuleEntries.length ||
+      filteredRuleEntries.every(([_, rules]) => !rules.length)) &&
+    !publication.rules?.length;
+  const oldRules = filteredRuleEntries.filter(([_, rules]) => rules.length);
+  const isNewRules = !!publication.rules?.length && !!publication.targetFolder;
+
+  return (
+    <>
+      {isNoRulesToDisplay && (
+        <p className="text-sm text-secondary" data-qa="availability-label">
+          {t(
+            'This publication will be available to all users in the organization',
+          )}
+        </p>
+      )}
+      {oldRules.map(([path, rules]) => (
+        <RuleListItem key={path} path={path} rules={rules} />
+      ))}
+      {isNewRules && (
+        <RuleListItem path={publication.targetFolder} rules={newRules} />
+      )}
+    </>
+  );
+}
